requirements.txt
requests
playwright
python-telegram-bot
beautifulsoup4
psycopg2